запустилось

+++ src/main/java/ru/skypro/homework/repository/AdsRepository.java
@@ -3,5 +3,7 @@ package ru.skypro.homework.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.skypro.homework.entity.Ad;
 
-public interface AdsRepository extends JpaRepository<Ad, Long> {
+public interface AdsRepository extends JpaRepository<Ad, Integer>
+{
+
 }
